Split things up properly
