+++ src/firmware/ths1206/pru0/pin_control.h
@@ -15,36 +15,15 @@
 #include <stdint.h>
 
 
-volatile register uint32_t __R30;
-volatile register uint32_t __R31;
-
 void set_pin( uint32_t bit );
-void set_pin( uint32_t bit )
-{
-   __R30 |= 1 << bit;
-}
 
 void clear_pin( uint32_t bit );
-void clear_pin( uint32_t bit )
-{
-   __R30 &= ~(1 << bit);
-}
 
 // TODO: Convert to 0/1?
 uint32_t read_pin( uint32_t bit );
-uint32_t read_pin( uint32_t bit )
-{
-   return __R31 & ~(1 << bit);
-}
 
-void write_reg( uint32_t value )
-{
-   __R30 = value;
-}
+void write_reg( uint32_t value );
 
-uint32_t read_reg()
-{
-   return __R31;
-}
+uint32_t read_reg();
 
 #endif /* PIN_CONTROL_H_ */
